feat #77 : 서비스 클래스 - 솔로 및 크루로 나누기
